requirements.txt
cairocffi==0.6
CairoSVG==2.0.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cssselect==1.0.1
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-easy-pdf==0.1.1
django-widget-tweaks==1.4.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.5
lxml==3.8.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
pip-autoremove==0.9.0
psycopg2==2.7.2
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
tinycss==0.4
tinycss2==0.5
urllib3==1.21.1
WeasyPrint==0.39
webencodings==0.5.1
xhtml2pdf==0.2b1
